requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
biopython>=1.79
torch>=1.9.0
requests>=2.26.0
pytest>=6.2.5
black>=21.6b0
python-dotenv>=0.19.0